+++ packages-exp/auth-exp/src/api/authentication/sign_up.ts
@@ -17,7 +17,7 @@
 
 import { Endpoint, HttpMethod, _performSignInRequest } from '..';
 import { IdTokenResponse } from '../../model/id_token';
-import { AuthCore } from '../../model/auth';
+import { Auth } from '@firebase/auth-types-exp';
 
 export interface SignUpRequest {
   returnSecureToken?: boolean;
@@ -31,7 +31,7 @@ export interface SignUpResponse extends IdTokenResponse {
 }
 
 export async function signUp(
-  auth: AuthCore,
+  auth: Auth,
   request: SignUpRequest
 ): Promise<SignUpResponse> {
   return _performSignInRequest<SignUpRequest, SignUpResponse>(