+++ packages/firebase/index.d.ts
@@ -1799,10 +1799,22 @@ declare namespace firebase.functions {
    */
   export class Functions {
     private constructor();
+
+    /**
+     * Modify this instance to communicate with the Cloud Functions emulator.
+     *
+     * Note: this must be called before this instance has been used to do any operations.
+     *
+     * @param host The emulator host (ex: localhost)
+     * @param port The emulator port (ex: 5001)
+     */
+    useEmulator(host: string, port: number): void;
+
     /**
      * Changes this instance to point to a Cloud Functions emulator running
      * locally. See https://firebase.google.com/docs/functions/local-emulator
      *
+     * @deprecated Prefer the useEmulator(host, port) method.
      * @param origin The origin of the local emulator, such as
      * "http://localhost:5005".
      */
@@ -3119,6 +3131,14 @@ declare namespace firebase.auth {
      * Sets the current language to the default device/browser preference.
      */
     useDeviceLanguage(): void;
+    /**
+     * Modify this Auth instance to communicate with the Firebase Auth emulator.  This must be
+     * called synchronously immediately following the first call to `firebase.auth()`.  Do not use
+     * with production credentials as emulator traffic is not encrypted.
+     *
+     * @param url The URL at which the emulator is running (eg, 'http://localhost:9099')
+     */
+    useEmulator(url: string): void;
     /**
      * Checks a password reset code sent to the user by email or other out-of-band
      * mechanism.
@@ -7396,7 +7416,7 @@ declare namespace firebase.storage {
     prefixes: Reference[];
     /**
      * Objects in this directory.
-     * You can call getMetadate() and getDownloadUrl() on them.
+     * You can call getMetadata() and getDownloadUrl() on them.
      */
     items: Reference[];
     /**
@@ -7826,40 +7846,6 @@ declare namespace firebase.firestore {
     ssl?: boolean;
 
     /**
-     * Specifies whether to use `Timestamp` objects for timestamp fields in
-     * `DocumentSnapshot`s. This is enabled by default and should not be
-     * disabled.
-     *
-     * Previously, Firestore returned timestamp fields as `Date` but `Date`
-     * only supports millisecond precision, which leads to truncation and
-     * causes unexpected behavior when using a timestamp from a snapshot as a
-     * part of a subsequent query.
-     *
-     * Now, Firestore returns `Timestamp` values for all timestamp values stored
-     * in Cloud Firestore instead of system `Date` objects, avoiding this kind
-     * of problem. Consequently, you must update your code to handle `Timestamp`
-     * objects instead of `Date` objects.
-     *
-     * If you want to **temporarily** opt into the old behavior of returning
-     * `Date` objects, you may **temporarily** set `timestampsInSnapshots` to
-     * false. Opting into this behavior will no longer be possible in the next
-     * major release of Firestore, after which code that expects Date objects
-     * **will break**.
-     *
-     * @example **Using Date objects in Firestore.**
-     * // With deprecated setting `timestampsInSnapshot: true`:
-     * const date : Date = snapshot.get('created_at');
-     * // With new default behavior:
-     * const timestamp : Timestamp = snapshot.get('created_at');
-     * const date : Date = timestamp.toDate();
-     *
-     * @deprecated This setting will be removed in a future release. You should
-     * update your code to expect `Timestamp` objects and stop using the
-     * `timestampsInSnapshots` setting.
-     */
-    timestampsInSnapshots?: boolean;
-
-    /**
      * An approximate cache size threshold for the on-disk data. If the cache grows beyond this
      * size, Firestore will start removing data that hasn't been recently used. The size is not a
      * guarantee that the cache will stay below that size, only that if the cache exceeds the given
@@ -7879,14 +7865,26 @@ declare namespace firebase.firestore {
      * buffer traffic indefinitely. Use of this option will cause some
      * performance degradation though.
      *
-     * This setting may be removed in a future release. If you find yourself
-     * using it to work around a specific network reliability issue, please
-     * tell us about it in
-     * https://github.com/firebase/firebase-js-sdk/issues/1674.
+     * This setting cannot be used with `experimentalAutoDetectLongPolling` and
+     * may be removed in a future release. If you find yourself using it to
+     * work around a specific network reliability issue, please tell us about
+     * it in https://github.com/firebase/firebase-js-sdk/issues/1674.
      *
      * @webonly
      */
     experimentalForceLongPolling?: boolean;
+
+    /**
+     * Configures the SDK's underlying transport (WebChannel) to automatically detect if
+     * long-polling should be used. This is very similar to `experimentalForceLongPolling`,
+     * but only uses long-polling if required.
+     *
+     * This setting will likely be enabled by default in future releases and cannot be
+     * combined with `experimentalForceLongPolling`.
+     *
+     * @webonly
+     */
+    experimentalAutoDetectLongPolling?: boolean;
 
     /**
      * Whether to skip nested properties that are set to `undefined` during
@@ -7921,17 +7919,6 @@ declare namespace firebase.firestore {
      * in all but the first tab.
      */
     synchronizeTabs?: boolean;
-
-    /**
-     * Whether to synchronize the in-memory state of multiple tabs. Setting this
-     * to `true` in all open tabs enables shared access to local persistence,
-     * shared execution of queries and latency-compensated local document updates
-     * across all connected instances.
-     *
-     * @deprecated This setting is deprecated. To enable synchronization between
-     * multiple tabs, please use `synchronizeTabs: true` instead.
-     */
-    experimentalTabSynchronization?: boolean;
 
     /**
      * Whether to force enable persistence for the client. This cannot be used
@@ -8042,6 +8029,16 @@ declare namespace firebase.firestore {
      * @param settings The settings to use.
      */
     settings(settings: Settings): void;
+
+    /**
+     * Modify this instance to communicate with the Cloud Firestore emulator.
+     *
+     * <p>Note: this must be called before this instance has been used to do any operations.
+     *
+     * @param host the emulator host (ex: localhost).
+     * @param port the emulator port (ex: 9000).
+     */
+    useEmulator(host: string, port: number): void;
 
     /**
      * Attempts to enable persistent storage, if possible.