+++ packages/firestore/exp/test/shim.ts
@@ -101,6 +101,10 @@ export class FirebaseFirestore
     initializeFirestore(this.app._delegate, settings);
   }
 
+  useEmulator(host: string, port: number): void {
+    this.settings({ host: `${host}:${port}`, ssl: false, merge: true });
+  }
+
   enablePersistence(settings?: legacy.PersistenceSettings): Promise<void> {
     return settings?.synchronizeTabs
       ? enableMultiTabIndexedDbPersistence(this._delegate)