+++ packages/firestore/test/unit/specs/spec_test_components.ts
@@ -143,7 +143,7 @@ export class MockMultiTabOfflineComponentProvider extends MultiTabOfflineCompone
     private readonly document: FakeDocument,
     onlineComponentProvider: OnlineComponentProvider
   ) {
-    super(onlineComponentProvider);
+    super(onlineComponentProvider, /* cacheSizeBytes= */ undefined);
   }
 
   createGarbageCollectionScheduler(
@@ -167,11 +167,6 @@ export class MockMultiTabOfflineComponentProvider extends MultiTabOfflineCompone
   }
 
   createPersistence(cfg: ComponentConfiguration): MockIndexedDbPersistence {
-    debugAssert(
-      cfg.persistenceSettings.durable,
-      'Can only start durable persistence'
-    );
-
     const persistenceKey = indexedDbStoragePrefix(
       cfg.databaseInfo.databaseId,
       cfg.databaseInfo.persistenceKey
@@ -182,13 +177,13 @@ export class MockMultiTabOfflineComponentProvider extends MultiTabOfflineCompone
       /* allowTabSynchronization= */ true,
       persistenceKey,
       cfg.clientId,
-      LruParams.withCacheSize(cfg.persistenceSettings.cacheSizeBytes),
+      LruParams.DEFAULT,
       cfg.asyncQueue,
       this.window,
       this.document,
       serializer,
       this.sharedClientState,
-      cfg.persistenceSettings.forceOwningTab
+      /* forceOwningTab= */ false
     );
   }
 }
@@ -208,10 +203,6 @@ export class MockMemoryOfflineComponentProvider extends MemoryOfflineComponentPr
   }
 
   createPersistence(cfg: ComponentConfiguration): Persistence {
-    debugAssert(
-      !cfg.persistenceSettings.durable,
-      'Can only start memory persistence'
-    );
     return new MockMemoryPersistence(
       this.gcEnabled
         ? MemoryEagerDelegate.factory